requirements.txt
streamlit 
langchain
langchain-groq
langchain-chroma
langchain-community
langchain-huggingface
langgraph
tavily-python
sentence-transformers
chromadb
pypdf
python-dotenv
